browser-use-agent/requirements.txt
browser-use
python-dotenv
fastapi
uvicorn
